bug fixes for streamtape, added some extra required libs, updated help handler etc

+++ handlers/streamdl.py
@@ -11,9 +11,17 @@ def exec_streamtapedl(client, message):
     if text == "":
         client.send_message(chat_id=chat_id, text="Enter a Streamtape Link")
     else:
-        helpers.streamtape_d.streamtape_dl(text)
-        print("Uploading...")
-        client.send_video(chat_id=chat_id, video = open("video.mp4","rb"))
-        print("Sent")
-        directory = os.getcwd()
-        os.remove(directory + "/video.mp4")
+        
+        
+        result = helpers.streamtape_d.streamtape_dl(text)
+        if result is True:
+            print("Uploading...")
+            client.send_message(chat_id = chat_id, text = "Uploading...", reply_to_message_id = msg_id)
+            client.send_video(chat_id=chat_id, video = open("video.mp4","rb"))
+            print("Sent")
+            directory = os.getcwd()
+            os.remove(directory + "/video.mp4")
+            print("video is cleaned")
+        else:
+           client.send_message(chat_id = chat_id, text = "Result not found", reply_to_message_id = msg_id) 
+
